requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langsmith
langchain_openai
langchain_google_genai
langchain_huggingface
torch
einops
sentence-transformers
langchain_nvidia_ai_endpoints
langchain_community
langchain_core
langchain_pymupdf4llm
langchain_text_splitters
pymupdf
pillow
pytesseract
python-pptx
python-docx
openpyxl
faiss-cpu
pyngrok
beautifulsoup4
requests
python-dotenv
markitdown[all]
pydantic
typing-extensions
uuid6